kinoz_fun/requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
django-debug-toolbar==4.4.2
idna==3.7
requests==2.32.3
sqlparse==0.5.0
urllib3==2.2.1
